+++ docs/assets/vue/components/fonts-card.js
@@ -1,13 +1,13 @@
+import themes from '../../data/themes.js';
 import PageCard from './page-card.js';
 import { defaultTitle, pairings, sameAs } from '/assets/data/fonts.js';
 import { themeConfig } from '/assets/data/theming.js';
 import { cssImport, getThemeCode } from '/assets/scripts/tweak/code.js';
-import themes from '/docs/themes/data.js';
 
 const template = `
   <page-card class="fonts-card" :info="computedPairing">
     <template #icon>
-      <wa-scoped slot="header" class="fonts-icon-host" inert>
+      <wa-scoped slot="header" class="fonts-icon-host" inert :key="html">
         <template v-html="html"></template>
         <template>
           <link rel="stylesheet" href="/dist/styles/native/content.css">