+++ docs/assets/vue/components/info-tip.js
@@ -2,7 +2,7 @@ const template = `
 	<slot>
   		<wa-icon :slot class="info-tip-default-trigger" :id="id" name="circle-question" variant="regular" tabindex="0"></wa-icon>
 	</slot>
-	<wa-tooltip :slot :for="id" ref="tooltip"><slot name="content"></slot></wa-tooltip>
+	<wa-tooltip :slot :for="id" ref="tooltip"><slot name="content">{{ text }}</slot></wa-tooltip>
   `;
 
 let maxUid = 0;
@@ -10,6 +10,7 @@ let maxUid = 0;
 export default {
   props: {
     slot: String,
+    text: String,
   },
   data() {
     let uid = ++maxUid;