+++ docs/assets/vue/components/palette-card.js
@@ -1,6 +1,6 @@
+import palettes from '../../data/palettes.js';
 import PageCard from './page-card.js';
 import { hues } from '/assets/data/index.js';
-import palettes from '/docs/palettes/data.js';
 
 // TODO import from data.js once available
 const allHues = [...hues, 'gray'];